9497 - add assign work item to self and also generisized docketClerkVerifies tests to pass in box and queue [skip ci]
